Ajustement code par rapp à nvl conception

+++ bibli/src/main/java/com/biblio/repository/ExemplaireRepository.java
@@ -3,5 +3,5 @@ package com.biblio.repository;
 import com.biblio.model.Exemplaire;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExemplaireRepository extends JpaRepository<Exemplaire, Long> {
+public interface ExemplaireRepository extends JpaRepository<Exemplaire, Integer> {
 }
